Disable debug mode
Significantly lowers the number of things being published to
NetworkTables which should improve performance. This is being done as a
potential fix for the LimeLight problems, since all of this
NetworkTables activity has the potential to overload the NetworkTables
thread, but it doesn't appear to have actually fixed the problem.